# src/day_57/internal/controllers/order.controller.go
/*
generated by comer,https://github.com/imoowi/comer
Copyright © 2023 jun<simpleyuan@gmail.com>
*/
package controllers

import (
	"codee_jun/internal/interfaces"
	"codee_jun/internal/models"
	"codee_jun/internal/services"
	"codee_jun/internal/utils/response"
	"errors"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/gin-gonic/gin/binding"
	"github.com/spf13/cast"
	"gorm.io/gorm"
)

// @Summary	分页列表(pagelist)
// @Tags	order
// @Accept	application/json
// @Produce	application/json
// @Param	Authorization	header		string				true	"Bearer 用户令牌"
// @Param   {object}             query                 models.OrderFilter  false "query参数"
// @Success 200                        {object} response.PageListT[models.Order] "成功"
// @Failure 400                        "请求错误"
// @Failure 401                        "token验证失败"
// @Failure 500                         "内部错误"
// @Router		/api/orders [get]
func OrderPageList(c *gin.Context) {
	var filter interfaces.IFilter = &models.OrderFilter{}
	err := c.ShouldBindQuery(&filter)
	if err != nil {
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}

	if 0 >= filter.GetPage() { //如果不传Page，默认为1
		filter.SetPage(1)
	}
	if 0 >= filter.GetPageSize() { //如果不传PageSize，默认取20条
		filter.SetPageSize(20)
	}
	if filter.GetPageSize() > 1000 {
		response.Error(c, `每一页不能超过1000条记录`, http.StatusBadRequest)
		return
	}
	result, err := services.Order.PageList(c, &filter)
	if err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			response.Error(c, err.Error(), http.StatusNotFound)
			return
		}
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}
	response.OK(c, result)
}

// @Summary	详情(one)
// @Tags		order
// @Accept		application/json
// @Produce	application/json
// @Param		Authorization	header	string	true	"Bearer 用户令牌"
// @Param		id				path	int		true	"id"
// @Success	200 						{object} models.Order "成功"
// @Failure 400                        "请求错误"
// @Failure 401                        "token验证失败"
// @Failure 500                         "内部错误"
// @Router		/api/orders/{id} [get]
func OrderOne(c *gin.Context) {
	id := c.Param(`id`)
	if id == `` {
		response.Error(c, `pls input id`, http.StatusBadRequest)
		return
	}

	one, err := services.Order.One(c, cast.ToUint(id))
	if err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			response.Error(c, err.Error(), http.StatusNotFound)
			return
		}
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}
	response.OK(c, one)
}

// @Summary	新增(add)
// @Tags		order
// @Accept		application/json
// @Produce	application/json
// @Param		Authorization	header	string	true	"Bearer 用户令牌"
// @Param 		{object} body  models.Order true "body"
// @Success	200
// @Failure 400                        "请求错误"
// @Failure 401                        "token验证失败"
// @Failure 500                         "内部错误"
// @Router		/api/orders [post]
func OrderAdd(c *gin.Context) {
	model := &models.Order{}
	err := c.ShouldBindBodyWith(&model, binding.JSON)
	if err != nil {
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}
	newId, err := services.Order.Add(c, model)
	if err != nil {
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}
	response.OK(c, newId)
}

// @Summary	更新(update)
// @Tags		order
// @Accept		application/json
// @Produce	application/json
// @Param		Authorization	header	string	true	"Bearer 用户令牌"
// @Param		id				path	int		true	"id"
// @Param 		{object} body  models.Order true "body"
// @Success	200
// @Failure 400                        "请求错误"
// @Failure 401                        "token验证失败"
// @Failure 500                         "内部错误"
// @Router		/api/orders/{id} [put]
func OrderUpdate(c *gin.Context) {
	id := c.Param(`id`)
	if id == `` {
		response.Error(c, `pls input id`, http.StatusBadRequest)
		return
	}
	model := make(map[string]any)
	err := c.ShouldBindBodyWith(&model, binding.JSON)
	if err != nil {
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}
	updated, err := services.Order.Update(c, model, cast.ToUint(id))
	if err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			response.Error(c, err.Error(), http.StatusNotFound)
			return
		}
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}
	response.OK(c, updated)
}

// @Summary	删除(delete)
// @Tags		order
// @Accept		application/json
// @Produce	application/json
// @Param		Authorization	header	string	true	"Bearer 用户令牌"
// @Param		id				path	int		true	"id"
// @Success	200
// @Failure 400                        "请求错误"
// @Failure 401                        "token验证失败"
// @Failure 500                         "内部错误"
// @Router		/api/orders/{id} [delete]
func OrderDel(c *gin.Context) {
	id := c.Param(`id`)
	if id == `` {
		response.Error(c, `pls input id`, http.StatusBadRequest)
		return
	}
	deleted, err := services.Order.Delete(c, cast.ToUint(id))
	if err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			response.Error(c, err.Error(), http.StatusNotFound)
			return
		}
		response.Error(c, err.Error(), http.StatusBadRequest)
		return
	}
	response.OK(c, deleted)
}