# src/day_57/internal/models/order.model.go
/*
generated by comer,https://github.com/imoowi/comer
Copyright © 2023 jun<simpleyuan@gmail.com>
*/
package models

import "gorm.io/gorm"

type OrderBase struct {
	Name string `json:"name" form:"name" gorm:"column:name;type:varchar(30);not null;comment:名" `
}

// Order表
type Order struct {
	gorm.Model `swaggerignore:"true"`
	OrderBase
}

// IModel.GetID实现
func (m *Order) GetId() uint {
	return m.ID
}
func (m *Order) SetId(id uint) {
	m.ID = id
}

// IModel.TableName实现
func (m *Order) TableName() string {
	return `order` + `s`
}